requirements/base.txt
absl-py==2.2.1
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.10
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
bitarray==3.4.2
blinker==1.9.0
cachelib==0.13.0
camelot-py==1.0.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
cytoolz==1.0.1
datasets==3.6.0
deepface==0.0.93
dill==0.3.8
distro==1.9.0
Django==5.2.3
dnspython==2.7.0
docx==0.2.4
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==6.0.0
Flask-PyMongo==3.0.1
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.2
fsspec==2025.3.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
hexbytes==1.3.1
huggingface-hub==0.32.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.1
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.3.2
lz4==4.4.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgspec==0.19.0
mtcnn==1.0.0
multidict==6.4.4
multiprocess==0.70.16
namex==0.0.8
networkx==3.3
numpy==2.1.3
oauthlib==3.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.13.0
pyparsing==3.2.3
pypdf
PyPDF2
pypdfium2
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
pytz==2025.2
pyunormalize==16.0.0
# pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.0.0
rlp==4.1.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
sympy==1.13.3
tabula-py==2.10.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tokenizers==0.19.1
toolz==1.0.0
torch
torchaudio
torchvision
tqdm==4.67.1
transformers==4.40.0
tweepy==4.15.0
types-requests==2.32.0.20250602
typing-inspection==0.4.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
web3==7.12.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
python-decouple
djangorestframework
django-storages
psycopg2-binary